requirements.txt
# Core CrewAI Dependencies
crewai>=0.70.0
crewai-tools>=0.12.0

# AI/ML Dependencies
openai>=1.0.0
pydantic>=2.0.0

# Web Framework (Streamlit UI)
streamlit>=1.28.0

# PDF Generation
reportlab>=4.0.0

# Markdown Processing
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Google APIs Integration
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Environment Management
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Data Processing
pandas>=2.0.0
